// ODOT.ARMS.Web/ClientApp/src/app/project/models/event-raw.ts
export class EventRaw {
  eventSrc: string;
  primaryTypeId: number;
  secondaryTypeId: number;
  invoiceNumber: string;
  publicCommentTxt: string;
  privateCommentTxt: string;
  userId: string;
  beginDate: Date;
  endDate: Date;
  activeInd: string;
  eventId: string;
  docCnt: number;
  projectId: string;
}